Update caching logic
Since we store the state in a normalized way, we can't reference other
objects unless we store them all locally. Which is fine.

However, this introduces a weird issue where we cache the fetch status
of an object and can get caught in an infinite loop of loading. I'll
address this in a later commit.