Consolidate ChapterRow
This commit removes the ReaderChapterRow in favor of a single,
consistent ChapterRow component which handles all states. This
consistency also removes the need for weird intermediate data-structures
such as "unread maps" and manually computing if a chapter was new or
not. Now, we simply pass chapters and bookmarks around, and the end
components determine what to show.