Add smarter lazyloading image dimensions
Previously, we assumed a static size for all images being loaded. This
resulted in a bunch of relayout and page jumping because the default
size rarely aligned with the actual image size.

This PR introduces a smarter method, where the most common image
dimensions from the images that are already loaded gets used as the
placeholder image size. In some quick testing, it seems to work quite
well, even for all-over-the-place series like _Shingeki no Kyojin_.