Move to /feed and /library routes
Rather than including the collection slug in all the routes, this commit
starts a transition towards simply using localStorage. The code was
already halfway there. This actually cleans up a lot of logic, by
consolidating where the collectionSlug gets referenced. It also removes
the need to reference it in low-level link components, just to generate
a URL.