Bump dependencies, remove extra withRouter calls
Now that react-redux@6.0.0 uses the new React Context API, it doesn't
suffer from the same update blocking as before, and we can remove the
`withRouter` calls for connected components that are direct children of
a `Route`.

`withRouter` calls for components that aren't direct children of a
`Route` still need to be there.